requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
scipy>=1.10.0
matplotlib>=3.7.0
Pillow>=9.5.0
tqdm>=4.65.0

# 3D reconstruction and geometry
open3d>=0.17.0
plyfile>=0.9

# Configuration and utilities
PyYAML>=6.0

# DeepLSD dependencies
kornia>=0.7.0
einops>=0.6.0

# Optional for advanced features
scikit-learn>=1.3.0
scikit-image>=0.21.0
